Fix send系の位置を移動
